fix: 순환 참조
