+++ src/main/java/com/nexters/goalpanzi/schedule/SchedulerConfig.java
@@ -6,15 +6,14 @@ import lombok.extern.slf4j.Slf4j;
 import org.quartz.JobDetail;
 import org.quartz.Scheduler;
 import org.quartz.SchedulerException;
-import org.quartz.simpl.SimpleThreadPool;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Lazy;
 import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
 import org.springframework.scheduling.quartz.SchedulerFactoryBean;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
-import java.util.concurrent.ExecutorService;
 
 @Slf4j
 @RequiredArgsConstructor
@@ -22,8 +21,6 @@ import java.util.concurrent.ExecutorService;
 public class SchedulerConfig {
     private final Scheduler scheduler;
     private final List<CustomAutomationJob> jobList;
-
-    private static final String SCHEDULER_THREAD_POOL_EXECUTOR = "schedulerThreadPool";
 
     @PostConstruct
     public void start() {
@@ -41,24 +38,5 @@ public class SchedulerConfig {
         } catch (SchedulerException e) {
             log.error("SchedulerException", e);
         }
-    }
-
-    @Bean
-    public SchedulerFactoryBean schedulerFactory(
-            @Qualifier(SCHEDULER_THREAD_POOL_EXECUTOR) ThreadPoolTaskExecutor threadPoolTaskExecutor
-    ) {
-        SchedulerFactoryBean factory = new SchedulerFactoryBean();
-        factory.setTaskExecutor(threadPoolTaskExecutor);
-        return factory;
-    }
-
-    @Bean(SCHEDULER_THREAD_POOL_EXECUTOR)
-    public ThreadPoolTaskExecutor executor() {
-        final ThreadPoolTaskExecutor executor = new ThreadPoolTaskExecutor();
-        executor.setCorePoolSize(1);
-        executor.setMaxPoolSize(1);
-        executor.setThreadNamePrefix("executor");
-        executor.initialize();
-        return executor;
     }
 }
